requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
databases==0.9.0
distro==1.9.0
fastapi==0.115.12
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.44.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.72.1
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.1
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
pywebpush
python-docx==1.1.0
openpyxl==3.1.2